add key 'w' to show winner(s) from voting
Add key 'w' to display winner(s) from [BHNT voting app](https://github.com/robotnyc/nodejs-vote-by-mac).

Pressing 'w' will initiate the HTTP GET request in the background and only on success will the
results be shown on the matelight.

Uses `ofLoadURLAsync` to perform HTTP GET to retrieve the JSON list of winners

For example, ["1"] or ["1", "10"] if there's a tie.